% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocs.R
\docType{data}
\name{alignment}
\alias{alignment}
\title{Sample alignment of soluable epoxide hydrolase family}
\format{An alignment object read with \code{\link[seqinr]{read.alignment}} function from seqinr package.
\describe{
  \item{nb}{A numeric: number of sequences}
  \item{nam}{A vector of characters: names of the sequences}
  \item{seq}{A vector of characters: amino acid sequences}
}}
\value{
A character vector of length of the aligned sequence containing consesus sequence based on the input alignment
}
\description{
was performed on a dataset comprising 311 soluble epoxide hydrolase peptide orthologous sequences acquired from UniProtKB. The alignment was performed and edited with MUSCLE algorithm in JALVIEW, respectively.
}
\examples{
data("alignment")
alignment
}
\references{
MUSCLE: \url{https://bmcbioinformatics.biomedcentral.com/articles/10.1186/1471-2105-5-113}\cr
JALVIEW: \url{https://academic.oup.com/bioinformatics/article/25/9/1189/203460/Jalview-Version-2-a-multiple-sequence-alignment}
}
\keyword{data}
