% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conservationFuncs.R
\name{alignment2matrix}
\alias{alignment2matrix}
\title{Convert alignment object to a matrix}
\usage{
alignment2matrix(alignment)
}
\arguments{
\item{alignment}{data loaded with \code{\link[seqinr]{read.alignment}} function}
}
\value{
Aligned sequences matrix where number of rows equals to number of aligned sequences and number of columns equals to the length length of aligned sequences
}
\description{
The function loads alignment into matrix to facilitate a convenient data manipulation
}
\examples{
data("alignment")
alignment = delete_isoforms(alignment)
matrix=alignment2matrix(alignment)
}
\keyword{alignment}
\keyword{conversion}
\keyword{matrix}
