% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conservationFuncs.R
\name{barplotshow}
\alias{barplotshow}
\title{Creates barplot with amino acid variation on the specified position}
\usage{
barplotshow(position, AA_variation)
}
\arguments{
\item{position}{A number of column of alignment to be visualized}

\item{AA_variation}{A percentage frequency of amino acids in the alignment, calculated with \code{\link{calculate_AA_variation}} function}
}
\value{
This function produces a barchart
}
\description{
This function facilitates a visual inspection of multiple sequence alignment (MSA) position variablity.
}
\examples{
data("small_alignment")
position = 100
threshold = 0.01
var_aa = calculate_AA_variation(small_alignment,threshold)
barplotshow(position, var_aa)
}
\keyword{plot}
