% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BALCONYfunctions.R
\name{calculate_AA_variation}
\alias{calculate_AA_variation}
\title{Calculate AA variations on each position of the multiple sequence alignment}
\usage{
calculate_AA_variation(alignment, threshold = NULL, grouped = F,
  grouping_method = "substitution_matrix", weights = NULL,
  pseudo_counts = F)
}
\arguments{
\item{alignment}{The data loaded with \code{\link[seqinr]{read.alignment}} function}

\item{threshold}{(optional) A number in range 0-1. A of minimal frequency of occurences of amino acids at each position. Default: all the residues are visualized.}

\item{grouped}{(optional) A logical indicating if the grouping of amino acids should be applied. Default: FALSE}

\item{grouping_method}{(optional) A string which specifies the grouping method to be used. One of following: 'substitution_matrix', 'polarity', 'size', 'aromaticity'. Default: 'substitution_matrix'. Default: 'substitution_matrix' if grouped is TRUE.}

\item{weights}{(optional) A vector of length equal number of sequences in the alignment object with weights to overcome the taxonomic bias in the conservation analysis.}

\item{pseudo_counts}{(optional) A logical indicating if pseudo-counts should be added to the MSA. Pseudo-counts can be used only in non-group mode and without weights. Using these options with pseudo-counts will be suppressed. Default: FALSE}
}
\value{
Returns list of three matrices with tabelarized symbols of the most common AA in alignment column, percentage values for contributed AA and combined one.
\item{AA }{A matrix of AA on all alignment positions with decreasing frequencies in columns}
\item{per}{The percentage of AA frequencies corresponding to the $AA}
\item{matrix}{A combination of this two. The best suited element for visual inspection of the variability at each position}
}
\description{
This function calculates AA variations on each position of the alignment which may be further used for the conservativity study of the set of sequences in quiestio
}
\details{
The output consists of amino acids and their fractions on each position of alignment. Amino acids with occurence frequencies lower than the threshold of user's choice are excluded.
}
\examples{
data("small_alignment")
alignment = delete_isoforms(small_alignment)
threshold=10
grouped = FALSE
var_aa=calculate_AA_variation(small_alignment, threshold, grouped)
}
