% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conservationFuncs.R
\name{calculate_pseudo_counts}
\alias{calculate_pseudo_counts}
\title{Calculate pseudo counts for alignment}
\usage{
calculate_pseudo_counts(alignment, substitution_mtx = NULL)
}
\arguments{
\item{alignment}{data loaded with \code{\link[seqinr]{read.alignment}}}

\item{substitution_mtx}{Matrix with amino acids substitution frequencies. Default: GONNET}
}
\value{
Matrix with pseudo counts of size 21x number of alignment columns
}
\description{
This function calculates pseudo-counts (as shown in \href{https://doi.org/10.1093/bioinformatics/12.2.135}{Henikoff et al. (1996)}) for an alignment with the use of substitution matrices. It is recommended to estimate amino acid frequencies for alignments with small number of sequences (in order to calculate reliable entropy scores)
}
\note{
Please note that when using other scoring matrix user needs to make sure that all alignment symbols are present there. Missing symbol will issue an error.
}
\examples{
data("alignment")
PC <- calculate_pseudo_counts(alignment)
}
\references{
Henikoff et al.(1996) Using substitution probabilities to improve position-specific scoring matrices, Bioinformatics, 12, 135–143\cr
Claverie (1994) Some useful statistical properties of position-weight matrices.
Comput. Chem., 18, 287-293
}
