% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BALCONYfunctions.R
\name{cons2seqs_ident}
\alias{cons2seqs_ident}
\title{Calculate identity of each sequence in the alignment to the consensus sequence.}
\usage{
cons2seqs_ident(alignment, consensus_seq)
}
\arguments{
\item{alignment}{Data loaded with \code{\link[seqinr]{read.alignment}} function}

\item{consensus_seq}{Consensus sequence (output of  \code{\link{consensus}} function)}
}
\value{
Numeric vector of identity score (percentage); positions in the numeric vector correspond to sequences in alignment, respectively
}
\description{
The function calculates identity of consensus to each sequence in the alignment. It facilitates an assessment of consensus accuracy and identification of outlying sequences in the alignment. Also, it can be used to weight conservativity metrics results in further steps of analysis with BALCONY package.
}
\details{
Returned values are percentage of identical symbols (AA and "-") in consensus sequence and aligned sequence.
}
\examples{
data("alignment")
alignment = delete_isoforms(alignment)
threshold=60
consensus=consensus(alignment, threshold)
true_consensus=cons2seqs_ident(alignment, consensus)
}
\keyword{consensus}
