% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BALCONYfunctions.R
\name{consensus}
\alias{consensus}
\title{Determine consensus sequence}
\usage{
consensus(alignment, threshold)
}
\arguments{
\item{alignment}{output of of \code{\link[seqinr]{read.alignment}} function or grouped alignment created with: \code{\link{align_seq_mtx2grs}} and \code{\link{alignment2matrix}}}

\item{threshold}{minimal fraction of amino acids on the certain position in all sequences of the alignment to be taken for consensus letter on this position; number in range 0-100.}
}
\value{
A character vector of length of the aligned sequence containing consesus sequence based on the input alignment
}
\description{
Function calculates consensus sequence for given alignment with a threshold of user's choice.
}
\details{
If maximum fraction of any amino acid on the certain position is lower than a threshold then "*" is printed instead.
}
\note{
Please note that this function masks the seqinr package function  \code{\link[seqinr]{consensus}}
}
\examples{
data("alignment")
alignment = delete_isoforms(alignment)
threshold=80 # Set the consensus threshold
consensus_sequence=consensus(alignment, threshold)
}
\keyword{consensus}
