% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conservationFuncs.R
\name{convert_AA_symbol}
\alias{convert_AA_symbol}
\title{Amino acids symbols conversion}
\usage{
convert_AA_symbol(amino_acids)
}
\arguments{
\item{amino_acids}{A character or vector of characters with amino acid(s) three letter code(s)}
}
\value{
A chracter or vector of characters with amino acids one letter code(s)
}
\description{
This function facilitates the conversion of three letter amino acids' codes to one letter equivalents.
}
\details{
In case a vector of amino acid three letter codes is provided the function returns a vector of their one letter equivalents.
}
\examples{
three_letter_codes = c("LEU", "VAL", "ALA")
convert_AA_symbol(three_letter_codes)
}
\keyword{amino_acids}
