% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conservationFuncs.R
\name{D_matrix}
\alias{D_matrix}
\title{Calculate substitution rate matrix between two amino acids}
\usage{
D_matrix(substitution_matrix)
}
\arguments{
\item{substitution_matrix}{A matrix with probablity of substitutions, e.g.  Gonnet substitution matrix}
}
\value{
A matrix of substitution probablities for all amino acids
}
\description{
This function is used to calculate Landgraf conservation metric. D_matrix contains substitution rates between two amino acids in the alignment, according to the following formula:\cr \cr
\deqn{D(a,b)= (d(a,a)-d(a,b))/d(a,a)}\cr
where:\cr
\eqn{d(a,a)} is a probability of AA substitution by itself\cr
\eqn{d(a,b)} is a probability of substitution of amino acid a with other amino acid.
}
\examples{
data("gonnet")
distance=D_matrix(gonnet)
}
