% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BALCONYfunctions.R
\name{delete_isoforms}
\alias{delete_isoforms}
\title{Delete protein isoforms from alignment object}
\usage{
delete_isoforms(alignment)
}
\arguments{
\item{alignment}{An object class alignment read with \code{\link[seqinr]{read.alignment}} function}
}
\value{
Alignment without isoforms
}
\description{
This function searches for isoforms in the alignment object (entries with "-digit|" in the name) and deletes them
}
\details{
The isoforms are detected as entries with \code{"-digit|"} in the sequence name. If no isoforms are detected this function prints a "No isiforms detected" notification instead
}
\examples{
data("alignment")
delete_isoforms(alignment)
}
\keyword{alignment}
