% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.apply_pseudocounts_position}
\alias{.apply_pseudocounts_position}
\title{apply the pseudocounts to the position}
\usage{
.apply_pseudocounts_position(alignment, df_table, position)
}
\arguments{
\item{alignment}{The data loaded with \code{\link[seqinr]{read.alignment}} function}

\item{df_table}{a data.frame object}

\item{position}{position id}
}
\value{
a matrix
}
\description{
apply the pseudocounts to the position
}
\examples{
#no example
}
