% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_consecutive_seq}
\alias{find_consecutive_seq}
\title{Find sequences of numbers in a numeric vector}
\usage{
find_consecutive_seq(vector)
}
\arguments{
\item{vector}{A numeric vector to be analyzed}
}
\value{
\item{values}{A vector of values starting the consecutive sequences}
\item{lengths}{A vector of lengths of identified sequences}
}
\description{
This function finds sequences of consecutive numbers in numeric vectors
}
\examples{
find_consecutive_seq(c(1,2,3,4,5,6,7,20,21,140,141,300,301,302))
}
