% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_seq}
\alias{find_seq}
\title{Get sequence by id in alignment.}
\usage{
find_seq(sequence_id, alignment)
}
\arguments{
\item{sequence_id}{identifier of desired sequence from alignment}

\item{alignment}{data loaded with \code{\link[seqinr]{read.alignment}}}
}
\value{
A string, the desired aligned sequence form alignment
}
\description{
This function allows to search for a sequence with its id. Useful for browsing a larg multiple sequence alignment data or for automatization purposes.
}
\examples{
data("alignment")
#creating library uniprot - PDB
lib=list(  c("Q84HB8","4I19","4QA9"),
           c("P34913","4JNC"),
           c("P34914","1EK2","1CR6","1EK1","1CQZ"))
sequence_id=find_seqid("1CQZ",lib)
sequence=find_seq(sequence_id, alignment)
}
