% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BALCONYfunctions.R
\name{get_remarks465_pdb}
\alias{get_remarks465_pdb}
\title{Get "REMARK 465" data from PDB file}
\usage{
get_remarks465_pdb(pdb_path, chain_identifier)
}
\arguments{
\item{pdb_path}{A string specifying the path tp the PDB file}

\item{chain_identifier}{A character specifying the chain to be considered}
}
\value{
\item{aa_numbers}{A numeric vector of indices of missing amino acids}
\item{chain}{A character specifying the chain which was considered in remark 465 data extraction}
}
\description{
This function extracts the data concerning missing amino acids in PDB protein structure from the PDB file
}
\examples{
require(Rpdb)
chain_identifier = "A"
pdb_path = system.file("extdata", "4jnc.pdb", package = "BALCONY")
print(pdb_path)
remark465_data = get_remarks465_pdb(pdb_path,chain_identifier)
}
