% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conservationFuncs.R
\name{get_seq_weights}
\alias{get_seq_weights}
\title{Get sequences weights}
\usage{
get_seq_weights(alignment)
}
\arguments{
\item{alignment}{data loaded with \code{\link[seqinr]{read.alignment}}}
}
\value{
A vector with weights of length equal to the number of sequences in the alignment
}
\description{
This function returns weights of the sequneces in the alignment object
}
\examples{
data("small_alignment")
alignment = small_alignment
weights = get_seq_weights(alignment)
}
\references{
Valdar, W. S. J. & Thornton, J. M. Protein–protein interfaces: Analysis of amino acid conservation in homodimers. Proteins: Structure, Function, and Bioinformatics 42, 108–124 (2001).
}
