% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocs.R
\docType{data}
\name{gonnet}
\alias{gonnet}
\title{Gonnet substitution matrix}
\format{An alignment object read with \code{\link[seqinr]{read.alignment}} function from seqinr package.
A data frame with 0 observations on the following 2 variables.
\describe{
  \item{AA names}{Names of amino acids included in the matrix}
  \item{matrix}{The substitution matrix itself}
}}
\source{
\url{http://imed.med.ucm.es/Tools/sias_help.html}
}
\value{
A character vector of length of the aligned sequence containing consesus sequence based on the input alignment
}
\description{
This dataset comprises the Gonnet substitution matrix which facilitates e.g. the calculation of Landgraf conservation score
}
\note{
Please note that this function masks the seqinr package function  \code{\link[seqinr]{consensus}}
}
\examples{
data("gonnet")
}
\keyword{data}
