% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BALCONYfunctions.R
\name{pairwise_alignment_MSA}
\alias{pairwise_alignment_MSA}
\title{Calculate pairwise alignment for whole MSA}
\usage{
pairwise_alignment_MSA(alignment)
}
\arguments{
\item{alignment}{An alignment object read with \code{\link[seqinr]{read.alignment}} function}
}
\value{
Matrix of alignment scores
}
\description{
For given alignment calculate pariwise alignments and returns alignment score.
}
\examples{
data("small_alignment")
\donttest{pairwiseAlignemnt_scores=pairwise_alignment_MSA(small_alignment)}
}
