% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structureFuncs.R
\name{prepare_structure_profile}
\alias{prepare_structure_profile}
\title{Combine the entropy data for structure building amino acids with their indices}
\usage{
prepare_structure_profile(structure, structure_entropy)
}
\arguments{
\item{structure}{A structure data, the output of \code{\link{create_structure_seq}} function}

\item{structure_entropy}{The entropy values for the structure building residues, the output of \code{\link{get_structures_entropy}} function}
}
\value{
Each element is a list of entropy values (matrix of entropy scores) and indices of residues building structure in protein of interest.
}
\description{
This function combines the entropy data for structure building amino acids with its indices. It prepares the data for convenient visualization or processing.
}
\examples{
data("alignment")
data("structure")
uniprot="P34914"
indices=get_structures_idx(structure)
protein_index = indices$proteinIndices
structure_index = indices$structureIndices
\donttest{entropy_scores_list=list(Schneider_entropy = schneider_conservativity(alignment),
                                   Escore_entropy = Escore_conservativity(alignment))
structure_entropy=get_structures_entropy(structure_index, entropy_scores_list)
structure_profile = prepare_structure_profile(structure, structure_entropy)}
}
\keyword{structure}
