% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structureFuncs.R
\name{RealValET_conservativity}
\alias{RealValET_conservativity}
\title{Calculate real-value Evolutionary Trace (ET)}
\usage{
RealValET_conservativity(alignment)
}
\arguments{
\item{alignment}{Data loaded with \code{\link[seqinr]{read.alignment}} function}
}
\value{
A vector of real valued ET score corresponding to each MSA column
}
\description{
This function allows to calculate real-valued ET for MSA.
}
\details{
Here, the real-valued ET is calculated using an evolutionary tree calculated for given alignment. The tree is calculated using UPGMA method. Real-valued ET score can be used as complimentary analysis of evolutionary entropy measures.
}
\examples{
data("small_alignment")
alignment = small_alignment
weights = get_seq_weights(alignment)
}
\references{
Mihalek, Res, Lichtarge, 2004
}
\keyword{conservation_metrics}
