% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocs.R
\docType{data}
\name{small_alignment}
\alias{small_alignment}
\title{Sample small alignment of soluable epoxide hydrolase family}
\format{An alignment object read with \code{\link[seqinr]{read.alignment}} function from seqinr package.
\describe{
  \item{nam}{A vector of characters: names of the sequences}
  \item{nb}{A numeric: number of sequences}
  \item{seq}{A vector of characters: amino acid sequences}
}}
\value{
A character vector of length of the aligned sequence containing consesus sequence based on the input alignment
}
\description{
This alignment consists of 10 proteins which belong to the soluable epoxide hydrolase family. The amino acid sequences were aligned using MUSCLE algorithm with default settings.
}
\details{
This is a smaller version of sample alignment which facilitates faster presentation of the functions capabilities.
}
\examples{
data("small_alignment")
small_alignment
}
\keyword{data}
