% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BALCONYfunctions.R
\name{substitution_mtx}
\alias{substitution_mtx}
\title{Read a substitution matrix}
\usage{
substitution_mtx(matrix_name)
}
\arguments{
\item{matrix_name}{A string with path to the substitution matrix in a text file to be read}
}
\value{
\item{names}{A vector of characters with amino acid names included in the matrix}
\item{matrix}{A numeric matrix with values}
}
\description{
This function facilitates reading of substitution matrices for further use
}
\examples{
path = system.file("extdata", "GONNET.txt", package = "BALCONY")
sub_mat = substitution_mtx(path)
}
