\name{gonnet}
\alias{gonnet}
\docType{data}
\title{
Gonnet substitution matrix
}
\description{
This dataset comprises the Gonnet substitution matrix which facilitates e.g. the calculation of Landgraf conservation score
}
\usage{data("gonnet")}
\format{
  A data frame with 0 observations on the following 2 variables.
  \describe{
    \item{AA names}{Names of amino acids included in the matrix}
    \item{matrix}{The substitution matrix itself}
  }
}
\source{
\url{http://imed.med.ucm.es/Tools/sias_help.html}
}

\examples{
data("gonnet")
}
\keyword{datasets}
